support build with local, fix reader, fix commands

+++ pkg/cmd/build/build.go
@@ -1,6 +1,7 @@
 package build
 
 import (
+	"bytes"
 	"context"
 	"fmt"
 	"io"
@@ -9,6 +10,8 @@ import (
 	"os/exec"
 	"path/filepath"
 	"strings"
+
+	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 
 	"github.com/solo-io/wasme/pkg/config"
 	"github.com/solo-io/wasme/pkg/model"
@@ -25,6 +28,7 @@ var log = logrus.StandardLogger()
 type buildOptions struct {
 	sourceDir    string
 	configFile   string
+	wasmFile     string
 	tag          string
 	storageDir   string
 	builderImage string
@@ -38,23 +42,26 @@ func BuildCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "build SOURCE_DIRECTORY [-b <bazel target>] [-t <name:tag>]",
 		Short: "Build a wasm image from the filter source directory using Bazel-in-Docker",
-		Args:  cobra.MinimumNArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			if len(args) == 0 {
-				return fmt.Errorf("invalid number of arguments")
+				if opts.wasmFile == "" {
+					return fmt.Errorf("must provide either SOURCE_DIRECTORY or --wasm-file to build an image")
+				}
+			} else {
+				opts.sourceDir = args[0]
 			}
-			opts.sourceDir = args[0]
 			return runBuild(opts)
 		},
 	}
 
 	cmd.Flags().StringVarP(&opts.tag, "tag", "t", "", "The image ref with which to tag this image. Specified in the format <name:tag>. Required")
 	cmd.Flags().StringVarP(&opts.configFile, "config", "c", "", "The path to the filter configuration file for the image. If not specified, defaults to <SOURCE_DIRECTOR>/filter-config.json. This file must be present in order to build the image.")
+	cmd.Flags().StringVarP(&opts.wasmFile, "wasm-file", "", "", "If specified, wasme will use the provided path to a compiled filter wasm to produce the image. The bazel build will be skipped and the wasm-file will be used instead.")
 	cmd.Flags().StringVar(&opts.storageDir, "store", "", "Set the path to the local storage directory for wasm images. Defaults to $HOME/.wasme/store")
 	cmd.Flags().StringVarP(&opts.builderImage, "image", "i", "quay.io/solo-io/ee-builder:"+version.Version, "Name of the docker image containing the Bazel run instructions. Modify to run a custom builder image")
 	cmd.Flags().StringVarP(&opts.buildDir, "build-dir", "b", ".", "Directory containing the target BUILD file.")
 	cmd.Flags().StringVarP(&opts.bazelOutput, "bazel-ouptut", "f", "filter.wasm", "Path relative to `bazel-bin` to the wasm file produced by running the Bazel target.")
-	cmd.Flags().StringVarP(&opts.bazelTarget, "bazel-target", "t", ":filter.wasm", "Name of the bazel target to run.")
+	cmd.Flags().StringVarP(&opts.bazelTarget, "bazel-target", "g", ":filter.wasm", "Name of the bazel target to run.")
 	return cmd
 }
 
@@ -74,51 +81,57 @@ func runBuild(opts buildOptions) error {
 		return err
 	}
 
-	sourceDir, err := filepath.Abs(opts.sourceDir)
-	if err != nil {
-		return err
-	}
+	var filterFile string
+	if opts.wasmFile != "" {
+		filterFile = opts.wasmFile
+	} else {
+		sourceDir, err := filepath.Abs(opts.sourceDir)
+		if err != nil {
+			return err
+		}
 
-	tmpDir, err := ioutil.TempDir("/tmp", "wasme")
-	if err != nil {
-		return err
-	}
-	defer os.RemoveAll(tmpDir)
-	// container paths are currently hard-coded in builder image
-	args := []string{
-		"run",
-		"--rm",
-		"-v", sourceDir + ":/src/workspace",
-		"-v", tmpDir + ":/tmp/build_output",
-		"-w", "/src/workspace",
-		"-e", "BUILD_BASE=" + opts.buildDir,
-		"-e", "BAZEL_OUTPUT=" + opts.bazelOutput,
-		"-e", "TARGET=" + opts.bazelTarget,
-		opts.builderImage,
+		tmpDir, err := ioutil.TempDir("/tmp", "wasme")
+		if err != nil {
+			return err
+		}
+		defer os.RemoveAll(tmpDir)
+
+		// container paths are currently hard-coded in builder image
+		args := []string{
+			"run",
+			"--rm",
+			"-v", sourceDir + ":/src/workspace",
+			"-v", tmpDir + ":/tmp/build_output",
+			"-w", "/src/workspace",
+			"-e", "BUILD_BASE=" + opts.buildDir,
+			"-e", "BAZEL_OUTPUT=" + opts.bazelOutput,
+			"-e", "TARGET=" + opts.bazelTarget,
+			opts.builderImage,
+		}
+
+		log.WithFields(logrus.Fields{
+			"args": args,
+		}).Info("running bazel-in-docker build...")
+
+		if err := docker(os.Stdout, os.Stderr, args...); err != nil {
+			return err
+		}
+
+		// filter.wasm currently hard-coded in bazel BUILD file
+		filterFile = filepath.Join(tmpDir, "filter.wasm")
 	}
 
 	log.WithFields(logrus.Fields{
-		"args": args,
-	}).Info("running bazel-in-docker build...")
-
-	if err := docker(os.Stdout, os.Stderr, args...); err != nil {
-		return err
-	}
-
-	// filter.wasm currently hard-coded in bazel BUILD file
-	tmpFile := filepath.Join(tmpDir, "filter.wasm")
-
-	log.WithFields(logrus.Fields{
-		"tmp_file": tmpFile,
-		"tag":      opts.tag,
+		"filter file": filterFile,
+		"tag":         opts.tag,
 	}).Info("adding image to cache...")
 
-	filterFile, err := os.Open(tmpFile)
-	if err != nil {
-		return err
+	fetchFilter := func() (model.Filter, error) {
+		return os.Open(filterFile)
 	}
 
-	descriptor, err := model.GetDescriptor(filterFile)
+	// need to read filter to generate descriptor
+	descriptor, err := getDescriptor(fetchFilter)
 	if err != nil {
 		return err
 	}
@@ -132,7 +145,7 @@ func runBuild(opts buildOptions) error {
 		return opts.tag + ":latest"
 	}()
 
-	image := store.NewStorableImage(imageRef, descriptor, filterFile, cfg)
+	image := store.NewStorableImage(imageRef, descriptor, fetchFilter, cfg)
 
 	if err := store.NewStore(opts.storageDir).Add(context.Background(), image); err != nil {
 		return err
@@ -155,4 +168,27 @@ func execCmd(stdout, stderr io.Writer, cmd string, args ...string) error {
 	command.Stderr = stderr
 	command.Stdout = stdout
 	return command.Run()
+}
+
+func getDescriptor(fetchFilter func() (model.Filter, error)) (ocispec.Descriptor, error) {
+	filter, err := fetchFilter()
+	if err != nil {
+		return ocispec.Descriptor{}, err
+	}
+	b, err := ioutil.ReadAll(filter)
+	if err != nil {
+		return ocispec.Descriptor{}, err
+	}
+	if closer, ok := filter.(io.ReadCloser); ok {
+		if err := closer.Close(); err != nil {
+			return ocispec.Descriptor{}, err
+		}
+	}
+
+	descriptor, err := model.GetDescriptor(bytes.NewBuffer(b))
+	if err != nil {
+		return ocispec.Descriptor{}, err
+	}
+
+	return descriptor, nil
 }
